Support for multiple auto-adjusted sign columns
With this change, having more than one sign on a line, and with the
'auto' setting on 'signcolumn', extra columns will shown automatically
to accomodate all the existing signs.

For example, suppose we have this view:

 5147             }
 5148
 5149             return sign->typenr;
 5150         }
 5151     }
 5152     return 0;
 5153 }
 5154

We have GitGutter installed, so it tells su about modified lines that
are not commmited. So let's change line 5152:

   5147             }
   5148
   5149             return sign->typenr;
   5150         }
   5151     }
 ~ 5152     return 0;
   5153 }
   5154

Now we add a mark over line 5152 using 'ma' in normal mode:

    5147             }
    5148
    5149             return sign->typenr;
    5150         }
    5151     }
a ~ 5152     return 0;
    5153 }
    5154

Previously, VIM and Neovim would have picked only one of the signs,
because there was no support for having multiple signs in a line.

Implements issue #990.

TBD: This feature requires more discussio and documentation.