Remove signs from deleted lines
Suppose we have highlights on a group of lines and we delete them:

 +     6 use std::ops::Deref;
 --+   7 use std::borrow::Cow;
 --+   8 use std::io::{Cursor};
       9 use proc_macro2::TokenStream;
      10 use syn::export::ToTokens;
 --+  11 use std::io::Write;
 >>   12 use std::ops::Deref;

Without this change, these signs will momentarily accumulate in
the sign column until the plugins wake up to refresh them.

+ --+ --+ --+ >>  6

Discussion: It may be better to extend the API a bit and allow this
to happen for only certain types of signs. For example, VIM marks
and vim-gitgutter removal signs may want to be presreved, unlike
line additions and linter highlights.